update song table create command
remove the limit of duration length when creating song table

+++ project1/sql_queries.py
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS songs (
     title VARCHAR(100),
     artist_id VARCHAR(20) NOT NULL,
     year INTEGER,
-    duration FLOAT(5)
+    duration FLOAT
 );
 """)
